# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LiveTrack real-time package tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Coordinates {
  lat: number;
  lng: number;
}

export interface RouteStop {
  name: string;
  lat: number;
  lng: number;
  done: boolean;
  active?: boolean;
  /** @nullable */
  timestamp?: string | null;
}

export interface Package {
  trackingId: string;
  status: string;
  origin: string;
  destination: string;
  carrier: string;
  estimatedDelivery: string;
  currentLat: number;
  currentLng: number;
  progressPct: number;
  currentLocationName?: string;
  weightLbs?: number;
  signatureRequired?: boolean;
}

export interface PackageRoute {
  trackingId: string;
  waypoints: Coordinates[];
  stops: RouteStop[];
}

export interface TrackingEvent {
  id: number;
  trackingId: string;
  message: string;
  location: string;
  timestamp: string;
}

export interface ViewerCount {
  trackingId: string;
  viewers: number;
}