# lib/api-zod/src/generated/types/coordinates.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LiveTrack real-time package tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface Coordinates {
  lat: number;
  lng: number;
}