# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LiveTrack real-time package tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface Package {
  trackingId: string;
  status: string;
  origin: string;
  destination: string;
  carrier: string;
  estimatedDelivery: string;
  currentLat: number;
  currentLng: number;
  progressPct: number;
  currentLocationName?: string;
  weightLbs?: number;
  signatureRequired?: boolean;
}