# lib/api-zod/src/generated/types/packageRoute.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LiveTrack real-time package tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { Coordinates } from './coordinates';
import type { RouteStop } from './routeStop';

export interface PackageRoute {
  trackingId: string;
  waypoints: Coordinates[];
  stops: RouteStop[];
}