# lib/api-zod/src/generated/types/routeStop.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LiveTrack real-time package tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface RouteStop {
  name: string;
  lat: number;
  lng: number;
  done: boolean;
  active?: boolean;
  /** @nullable */
  timestamp?: string | null;
}