# lib/api-zod/src/generated/types/trackingEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LiveTrack real-time package tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface TrackingEvent {
  id: number;
  trackingId: string;
  message: string;
  location: string;
  timestamp: string;
}